Change the default states we notify for
Was: finished,unconfirmed
Now: finished

Notifying for every unconfirmed run was overly spammy. It was leading to
alert fatigue where people were ignoring most notifications from the
bot.

Instead, only notify on run failures. In this way, the signal-to-noise
ratio should be extremely high.

Ideally, we would only like to be notified about unconfirmed runs that
have been forgotten. A heuristic might be "not confirmed within x
minutes". That would be a nice future feature...